renaming and doc cosmetics
